requirements.txt
amqp==5.1.1
asgiref==3.7.2
beautifulsoup4==4.12.2
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.3
Django==4.2.4
django-csp==3.7
django-prometheus==2.3.1
django-simple-history==3.3.0
django-sri==0.6.0
django-taggit==4.0.0
django-tinymce==3.6.1
idna==3.4
kombu==5.3.1
numpy==1.24.4
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
Pillow==10.0.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pycparser==2.21
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uWSGI==2.0.22
vine==5.0.0
wcwidth==0.2.6
